requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==44.0.3
Django==5.2.1
django-auth-adfs==1.15.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
iniconfig==2.1.0
ms-identity-web==0.0.4
msal==1.32.3
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
requests==2.32.3
sqlparse==0.4.4
tzdata==2025.2
urllib3==2.4.0
